Revert "refactor: migrate connector oauth routes to zero namespace (#7473)"
This reverts commit eae55012401322f975897a68097cab008a78b650.
