+++ turbo/apps/web/src/__tests__/api-test-helpers.ts
@@ -95,7 +95,7 @@ import { POST as storageCommitRoute } from "../../app/api/storages/commit/route"
 import { POST as setSecretRoute } from "../../app/api/zero/secrets/route";
 import { POST as setVariableRoute } from "../../app/api/zero/variables/route";
 
-import { GET as connectorCallbackRoute } from "../../app/api/zero/connectors/[type]/callback/route";
+import { GET as connectorCallbackRoute } from "../../app/api/connectors/[type]/callback/route";
 import { composeJobs } from "../db/schema/compose-job";
 import { connectors } from "../db/schema/connector";
 import { connectorSessions } from "../db/schema/connector-session";
@@ -1624,7 +1624,7 @@ export async function insertStalePendingRun(
  * Create a test connector via API routes.
  *
  * - api-token: calls POST /api/connectors/:type/token
- * - oauth: calls GET /api/zero/connectors/:type/callback with MSW mocks
+ * - oauth: calls GET /api/connectors/:type/callback with MSW mocks
  *
  * @param options - Connector configuration
  */
@@ -1816,7 +1816,7 @@ const OAUTH_PROVIDER_MOCKS: Record<
 };
 
 /**
- * Create an OAuth connector via GET /api/zero/connectors/:type/callback with MSW mocks.
+ * Create an OAuth connector via GET /api/connectors/:type/callback with MSW mocks.
  */
 async function createTestOAuthConnector(options?: {
   type?: ConnectorType;
@@ -1855,9 +1855,7 @@ async function createTestOAuthConnector(options?: {
 
   // Create callback request with proper cookies
   const state = "test-oauth-state";
-  const url = new URL(
-    `http://localhost:3000/api/zero/connectors/${type}/callback`,
-  );
+  const url = new URL(`http://localhost:3000/api/connectors/${type}/callback`);
   url.searchParams.set("code", "test-code");
   url.searchParams.set("state", state);
